feat(items): enhance item entry management with initial and current quantities, units, and container types

+++ backend/dotnet/src/Domu.Api/Features/Spaces/Application/Items/Contracts/ItemView.cs
@@ -8,7 +8,7 @@ public sealed record ItemView(
     string Name,
     string? Category,
     string? Barcode,
-    int TotalQuantity,
+    decimal TotalQuantity,
     IReadOnlyList<ItemEntryView> Entries)
 {
     public static ItemView FromDomain(Item item)
